// generated/src/app/api-models/db-cols-not-case-sensitive.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class DbColsNotCaseSensitive {
  caseInsensitiveTblNmPk: string;
  caseInsensitiveColNmPk: string;
  insertRowTms: Date;
  insertUserSysId: string;
  insertProcessSysId: string;
  updateRowTms: Date;
  updateUserSysId: string;
  updateProcessSysId: string;
}